pyproject.toml
[project]
name = "apple-sugar-detection"
version = "0.1.0"
description = "苹果糖度检测"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.2",
    "tensorboard>=2.20.0",
    "timm>=1.0.20",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]
